Refined plotter code;Added docstrings
