+++ ivy/__init__.py
@@ -1,3 +1,5 @@
-from . import train
 from . import loss
 from . import metrics
+from . import train
+
+__all__ = ["loss", "metrics", "train"]
