+++ ivy/metrics/__init__.py
@@ -1,1 +1,4 @@
-from .evaluator import Evaluator
+from . import plots
+from . import evaluator
+
+__all__ = ["plots", "evaluator"]
