+++ ivy/metrics/plots/__init__.py
@@ -1,0 +1,3 @@
+from ._confusion_matrix_plotter import ConfusionMatrixPlotter
+from ._roc_plotter import ROCPlotter
+from ._training_loss_plotter import TrainingLossPlotter
